Main: funciones 1-4 implementadas y revisadas

+++ betterio.h
@@ -8,6 +8,6 @@ void eliminatespace(char word[]);
 /*El numero entero ingresado entra a procesos de verificacion entre maximos y minimos, que no sean datos basuras o letras*/
 int ENTnumberverification(int min, int max);
 /*El numero decimal ingresado entra a procesos de verificacion entre maximos y minimos, que no sean datos basuras o letras*/
-float DECnumberverification(int min, int max);
+float DECnumberverification(float min, float max);
 
 #endif
